+++ orange_cb_recsys/utils/const.py
@@ -4,8 +4,6 @@ import logging
 home_path = str(Path.home())
 DEVELOPING = True
 
-logging.basicConfig(level=logging.INFO)
-
 logger = logging.getLogger('logger')
 
 logger.setLevel(logging.INFO)
